Implement single-post view
12/21: A user may click on an image post to open a separate screen with all of that post's info. Most notably, all images are loaded into the view and a user may scroll through each photo.

On iOS, a user may click into, and zoom in/out of, a photo (although this is harder on the emulator). Only clicking into a photo is allowed on Android (no zoom). On the web (dev purposes), clicking into a photo opens a new tab with just that photo.

This PR also includes proper loadup and routing of posts on the profile page, since that is related to single image-posts.

Comments may look nicer as a vertical continuation of the single post page? This is something that I could patch in the future, if it's of interest within our team. Currently, I'm using our same overlay for continuity.